+++ weblate/utils/search.py
@@ -196,7 +196,10 @@ def field_name(field, suffix="icontains"):
     if field in STRING_FIELD_MAP:
         return "{}__{}".format(STRING_FIELD_MAP[field], suffix)
     if field in EXACT_FIELD_MAP:
-        return "{}__iexact".format(EXACT_FIELD_MAP[field])
+        # Change contains to exact, do not change other (eg. regex)
+        if suffix == "icontains":
+            suffix = "iexact"
+        return "{}__{}".format(EXACT_FIELD_MAP[field], suffix)
     return field